done code jwt login

+++ src/main/java/com/projectJava/Hylife/User/Repository/UserProfileRepository.java
@@ -3,5 +3,5 @@ package com.projectJava.Hylife.User.Repository;
 import com.projectJava.Hylife.User.Entity.UserProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserProfileRepository extends JpaRepository<UserProfile,Long> {
+public interface UserProfileRepository extends JpaRepository<UserProfile,Integer> {
 }
